+++ lms/templates/dashboard.html
@@ -76,6 +76,7 @@ from django.core.urlresolvers import reverse
       % for dashboard_index, (course, enrollment) in enumerate(course_enrollment_pairs):
         <% show_courseware_link = (course.id in show_courseware_links_for) %>
         <% cert_status = cert_statuses.get(course.id) %>
+        <% credit_status = credit_statuses.get(course.id) %>
         <% show_email_settings = (course.id in show_email_settings_for) %>
         <% course_mode_info = all_course_modes.get(course.id) %>
         <% show_refund_option = (course.id in show_refund_option_for) %>
@@ -83,8 +84,7 @@ from django.core.urlresolvers import reverse
         <% is_course_blocked = (course.id in block_courses) %>
         <% course_verification_status = verification_status_by_course.get(course.id, {}) %>
         <% course_requirements = courses_requirements_not_met.get(course.id) %>
-        <% credit_message = credit_messages.get(unicode(course.id)) %>
-        <%include file='dashboard/_dashboard_course_listing.html' args="course=course, enrollment=enrollment, show_courseware_link=show_courseware_link, cert_status=cert_status, show_email_settings=show_email_settings, course_mode_info=course_mode_info, show_refund_option = show_refund_option, is_paid_course = is_paid_course, is_course_blocked = is_course_blocked, verification_status=course_verification_status, course_requirements=course_requirements, dashboard_index=dashboard_index, share_settings=share_settings, credit_message=credit_message, user=user" />
+        <%include file='dashboard/_dashboard_course_listing.html' args="course=course, enrollment=enrollment, show_courseware_link=show_courseware_link, cert_status=cert_status, credit_status=credit_status, show_email_settings=show_email_settings, course_mode_info=course_mode_info, show_refund_option = show_refund_option, is_paid_course = is_paid_course, is_course_blocked = is_course_blocked, verification_status=course_verification_status, course_requirements=course_requirements, dashboard_index=dashboard_index, share_settings=share_settings, user=user" />
       % endfor
 
       % if settings.FEATURES.get('CUSTOM_COURSES_EDX', False):
