--- lms/templates/dashboard/_dashboard_credit_information.html
@@ -1,70 +1,0 @@
-<%page args="credit_message" />
-
-<%!
-from django.utils.translation import ugettext as _
-from course_modes.models import CourseMode
-from util.date_utils import get_default_time_display
-%>
-<%namespace name='static' file='../static_content.html'/>
-
-<%block name="js_extra" args="credit_message">
-  <%static:js group='credit_wv'/>
-  <script type="text/javascript">
-    $(document).ready(function() {
-      $.ajaxSetup({
-        headers: {
-          'X-CSRFToken': $.cookie('csrftoken')
-        },
-        dataType: 'json'
-      });
-      $(".purchase-credit-btn").click(function() {
-        var data = {
-            user_id: "${credit_message['user_id']}",
-            course_id: "${credit_message['course_id']}"
-        };
-        Logger.log('edx.credit.shared', data);
-      });
-    });
-  </script>
-</%block>
-
-<div class="message message-status is-shown">
-    <p>
-        % if credit_message["status"] == "requirements_meet":
-            <span>
-            % if credit_message["urgent"]:
-                ${_("{username}, your eligibility for credit expires on {expiry}. Don't miss out!").format(
-                        username=credit_message["user_full_name"],
-                        expiry=get_default_time_display(credit_message["expiry"])
-                    )
-                }
-            % else:
-                ${_("{congrats} {username}, You have meet requirements for credit.").format(
-                        congrats="<b>Congratulations</b>",
-                        username=credit_message["user_full_name"]
-                    )
-                }
-            % endif
-            </span>
-            <span class="purchase_credit"> <a class="btn purchase-credit-btn" href="" target="_blank">${_("Purchase Credit")}</a> </span>
-
-        % elif credit_message["status"] == "pending":
-            ${_("Thank you, your payment is complete, your credit is processing. Please see {provider_link} for more information.").format(
-                    provider_link='<a href="#" target="_blank">{}</a>'.format(credit_message["provider"]["display_name"])
-                )
-            }
-        % elif credit_message["status"] == "approved":
-            ${_("Thank you, your credit is approved. Please see {provider_link} for more information.").format(
-                    provider_link='<a href="#" target="_blank">{}</a>'.format(credit_message["provider"]["display_name"])
-                )
-            }
-        % elif credit_message["status"] == "rejected":
-            ${_("Your credit has been denied. Please contact {provider_link} for more information.").format(
-                    provider_link='<a href="#" target="_blank">{}</a>'.format(credit_message["provider"]["display_name"])
-                )
-            }
-        % endif
-
-    </p>
-
-</div>
