+++ openedx/core/djangoapps/credit/api/__init__.py
@@ -1,0 +1,8 @@
+"""
+Credit Python API.
+
+This module aggregates the API functions from the eligibility and provider APIs.
+"""
+
+from .eligibility import *  # pylint: disable=wildcard-import
+from .provider import *  # pylint: disable=wildcard-import
