+++ openedx/core/djangoapps/credit/signals.py
@@ -39,24 +39,24 @@ def listen_for_grade_calculation(sender, username, grade_summary, course_key, de
         kwargs : None
 
     """
-    from openedx.core.djangoapps.credit.api import (
-        is_credit_course, get_credit_requirement, set_credit_requirement_status
-    )
+    # This needs to be imported here to avoid a circular dependency
+    # that can cause syncdb to fail.
+    from openedx.core.djangoapps.credit import api
 
     course_id = CourseKey.from_string(unicode(course_key))
-    is_credit = is_credit_course(course_id)
+    is_credit = api.is_credit_course(course_id)
     if is_credit:
-        requirement = get_credit_requirement(course_id, 'grade', 'grade')
-        if requirement:
-            criteria = requirement.get('criteria')
+        requirements = api.get_credit_requirements(course_id, namespace='grade')
+        if requirements:
+            criteria = requirements[0].get('criteria')
             if criteria:
                 min_grade = criteria.get('min_grade')
                 if grade_summary['percent'] >= min_grade:
                     reason_dict = {'final_grade': grade_summary['percent']}
-                    set_credit_requirement_status(
+                    api.set_credit_requirement_status(
                         username, course_id, 'grade', 'grade', status="satisfied", reason=reason_dict
                     )
                 elif deadline and deadline < timezone.now():
-                    set_credit_requirement_status(
+                    api.set_credit_requirement_status(
                         username, course_id, 'grade', 'grade', status="failed", reason={}
                     )
